added examples folder with Manifest and Project for reproducibility
