+++ examples/Project.toml
@@ -1,0 +1,15 @@
+[deps]
+BeliefUpdaters = "8bb6e9a1-7d73-552c-a44a-e5dc5634aac4"
+Cairo = "159f3aea-2a34-519c-b102-8c37f9878175"
+POMDPGifs = "7f35509c-0cb9-11e9-0708-2928828cdbb7"
+POMDPModelChecking = "abefb91b-a28c-5ab9-9bd9-026e532d7b0e"
+POMDPModelTools = "08074719-1b2a-587c-a292-00f91cc44415"
+POMDPModels = "355abbd5-f08e-5560-ac9e-8b5f2592a0ca"
+POMDPSimulators = "e0d0a172-29c6-5d4e-96d0-f262df5d01fd"
+POMDPs = "a93abf59-7444-517b-a68a-c42f96afdd7d"
+ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
+Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
+RockSample = "de008ff0-c357-11e8-3329-7fe746fe836e"
+SARSOP = "cef570c6-3a94-5604-96b7-1a5e143043f2"
+Spot = "f11abc24-ce50-11e8-2475-af6658d13f2b"
+Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
